+++ Tests/SwiftTileJSONTests/EncodingTests.swift
@@ -11,12 +11,15 @@ import Testing
 
 struct EncodingTests {
     @Test func extendedFieldsEncoding() {
-        let tileJSON = TileJSON(tiles: ["http://a.tileserver.org/{z}/{x}/{y}"])
-        let extendedFields: [String: ExtendedField] = [
-            "something_custom": "this is my unique field",
-            "another_custom": 42
-        ]
-        let extendedTileJSON = ExtendedTileJSON(tileJSON: tileJSON, extendedFields: extendedFields)
+        let extendedTileJSON = TileJSON(
+            TileJSON.Base(
+                tiles: ["http://a.tileserver.org/{z}/{x}/{y}"]
+            ),
+            extendedFields: [
+                "something_custom": "this is my unique field",
+                "another_custom": 42
+            ]
+        )
         
         let jsonData = try! JSONEncoder().encode(extendedTileJSON)
         let decoded = try! JSONSerialization.jsonObject(with: jsonData, options: []) as! [String: Any]